# go_api/types_gen.go
// Package main provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package main

// Error エラーオブジェクトの共通スキーマ
type Error struct {
	// Message エラーメッセージ
	Message *string `json:"message,omitempty"`

	// State エラーの種別
	State string `json:"state"`
}

// ToDo Todoオブジェクトの共通スキーマ
type ToDo struct {
	// AssaginPerson Todoを担当する人の名前
	AssaginPerson *string `json:"assagin_person"`

	// CreatedAt Todoが作成された時刻（UNIXタイムで単位は秒）
	CreatedAt *int64 `json:"created_at,omitempty"`

	// Description Todoの説明
	Description *string `json:"description,omitempty"`

	// Id ユニークID
	Id *int `json:"id,omitempty"`

	// IsComplete Todoの完了・未完了を示すフラグ（trueが完了）
	IsComplete *bool `json:"is_complete,omitempty"`

	// Title Todoのタイトル
	Title string `json:"title"`

	// UpdatedAt Todoが更新された時刻（UNIXタイムで単位は秒）
	UpdatedAt *int64 `json:"updated_at,omitempty"`
}

// PostTodoJSONBody defines parameters for PostTodo.
type PostTodoJSONBody struct {
	AssiginPerson string  `json:"assigin_person"`
	Description   *string `json:"description,omitempty"`

	// Title Todoのタイトル
	Title string `json:"title"`
}

// UpdateTodoIdJSONBody defines parameters for UpdateTodoId.
type UpdateTodoIdJSONBody struct {
	AssiginPerson string  `json:"assigin_person"`
	Description   *string `json:"description,omitempty"`

	// Title Todoのタイトル
	Title string `json:"title"`
}

// GetTodosParams defines parameters for GetTodos.
type GetTodosParams struct {
	// Offset 結果セットの収集を開始する前にスキップするアイテムの数
	Offset int `form:"offset" json:"offset"`

	// Limit 返却するアイテムの数
	Limit int `form:"limit" json:"limit"`
}

// PostTodoJSONRequestBody defines body for PostTodo for application/json ContentType.
type PostTodoJSONRequestBody PostTodoJSONBody

// UpdateTodoIdJSONRequestBody defines body for UpdateTodoId for application/json ContentType.
type UpdateTodoIdJSONRequestBody UpdateTodoIdJSONBody